Replace `NamespaceManager` with `NamespaceRegistrar`
